get real data and view it

+++ lib/application/providers/humidity_string_provider.dart
@@ -1,5 +1,5 @@
 abstract class HumidityStringProvider {
   const HumidityStringProvider();
 
-  String getString(double humidity);
+  String getString(int humidity);
 }
